cluster-autoscaler: Replace ambiguous "delete node" wording in FAQ
"delete node" has a specific meaning in the context of the Kubernetes
API: deleting the node object. However, the cluster-autoscaler never
does that; it terminates the underlying instance and expects the
[cloud-node-controller](https://kubernetes.io/docs/concepts/architecture/cloud-controller/#node-controller)
to remove the corresponding node from Kubernetes.

Replace all mentions of "deleting" a node with "terminating" to
disambiguate this.

Signed-off-by: Matthias Rampke <mr@soundcloud.com>