cloudprovider/digitalocean: fix unit tests
The `godo/util` package is not used and wasn't changed during vendoring
of the godo package. Seems like Travis didn't picked up this in the
initial PR. Let's remove the `util/` folder as it's not needed and also
add a documentation to list the changes we need to make in the `godo`
package.